# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Data Analyst Agent API
  version: 1.0.0
  description: |
    AI Data Analyst Agent 前后端共享契约。
    具体通用规则见 API_INTEGRATION_STANDARD.md。
  contact:
    name: AI Data Analyst Agent 项目组
servers:
  - url: http://localhost:8000/api/v1
    description: 本地开发
  - url: /api/v1
    description: 同源部署
...
security:
  - bearerAuth: []
paths:
...
  /projects/{project_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Projects]
      operationId: getProject
      summary: 获取项目详情
      responses:
        "200":
          description: 项目详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
    patch:
      tags: [Projects]
      operationId: updateProject
      summary: 更新项目元数据
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: 更新后的项目
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
    delete:
      tags: [Projects]
      operationId: archiveProject
      summary: 归档项目
      responses:
        "204":
          description: 已归档
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
...
  /projects/{project_id}/dataset-versions/{version_id}/quality-scans:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/VersionId"
    post:
      tags: [Quality]
      operationId: createQualityScan
      summary: 启动质量扫描
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QualityScanRequest"
      responses:
        "202":
          description: 质量扫描任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/dataset-versions/{version_id}/quality-issues:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/VersionId"
    get:
      tags: [Quality]
      operationId: listQualityIssues
      summary: 查询质量问题
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, accepted, resolved, ignored]
        - name: issue_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 质量问题列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityIssuePage"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/quality-issues/{issue_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/IssueId"
    patch:
      tags: [Quality]
      operationId: updateQualityIssue
      summary: 更新质量问题处置状态
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QualityIssueUpdate"
      responses:
        "200":
          description: 更新后的问题
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QualityIssue"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
...
  /projects/{project_id}/runs:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Runs]
      operationId: listRuns
      summary: 查询分析运行
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, blocked, succeeded, failed, cancelled]
        - name: dataset_version_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: 运行列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRunPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      tags: [Runs]
      operationId: createRun
      summary: 根据已确认 AnalysisSpec 启动运行
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisRunCreate"
      responses:
        "202":
          description: 分析任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/runs/{run_id}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/RunId"
    get:
      tags: [Runs]
      operationId: getRun
      summary: 获取运行详情与步骤
      responses:
        "200":
          description: 运行详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Error"
  /projects/{project_id}/runs/{run_id}/cancel:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
      - $ref: "#/components/parameters/RunId"
    post:
      tags: [Runs]
      operationId: cancelRun
      summary: 取消分析运行
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: 取消后的运行
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisRun"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
...
  /projects/{project_id}/reports:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      tags: [Reports]
      operationId: createReportExport
      summary: 创建报告、Notebook、数据或 Manifest 导出
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportExportRequest"
      responses:
        "202":
          description: 导出任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "422":
          $ref: "#/components/responses/ValidationError"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: datatrace_session
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^prj_[A-Za-z0-9]+$"
...
    VersionId:
      name: version_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^dsv_[A-Za-z0-9]+$"
    IssueId:
      name: issue_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^issue_[A-Za-z0-9]+$"
...
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
        pattern: "^run_[A-Za-z0-9]+$"
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 资源 revision，格式为双引号包围的整数
      schema:
        type: string
        example: '"3"'
  responses:
...
    Error:
      description: 错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: 资源不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 资源不存在或无权访问
              request_id: req_01JABC
              retryable: false
              details: {}
    Conflict:
      description: 当前状态或重复请求冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    PreconditionFailed:
      description: 客户端资源版本已过期
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VERSION_CONFLICT
              message: 资源已被修改，请刷新后重试
              request_id: req_01JABC
              retryable: false
              details:
                current_revision: 4
    ValidationError:
      description: 业务校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ProjectUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, "null"]
          maxLength: 2000
        timezone:
          type: string
        language:
          type: string
          enum: [zh-CN, en-US]
    Project:
      type: object
      required:
        - project_id
        - name
        - timezone
        - language
        - status
        - revision
        - created_at
        - updated_at
      properties:
        project_id:
          type: string
          example: prj_01JABC
        name:
          type: string
        description:
          type: [string, "null"]
        timezone:
          type: string
        language:
          type: string
        status:
          type: string
          enum: [active, archived]
        current_dataset_version_id:
          type: [string, "null"]
        revision:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    QualityScanRequest:
      type: object
      additionalProperties: false
      properties:
        analysis_spec_id:
          type: [string, "null"]
        rules:
          type: array
          items:
            type: string
        force:
          type: boolean
          default: false
    QualityIssue:
      type: object
      required:
        - issue_id
        - dataset_version_id
        - issue_type
        - severity
        - status
        - metrics
        - revision
        - created_at
      properties:
        issue_id:
          type: string
          example: issue_01JABC
        dataset_version_id:
          type: string
        issue_type:
          type: string
          example: missing_values
        column:
          type: [string, "null"]
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, accepted, resolved, ignored]
        title:
          type: string
        explanation:
          type: string
        metrics:
          type: object
          additionalProperties: true
        sample_rows:
          type: array
          items:
            type: object
            additionalProperties: true
        recommendation:
          type: [string, "null"]
        decision_reason:
          type: [string, "null"]
        revision:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
    QualityIssueUpdate:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [open, accepted, ignored]
        reason:
          type: [string, "null"]
          maxLength: 1000
    QualityIssuePage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/QualityIssue"
...
    AnalysisRunCreate:
      type: object
      additionalProperties: false
      required: [analysis_spec_id, dataset_version_id]
      properties:
        analysis_spec_id:
          type: string
        dataset_version_id:
          type: string
        run_kind:
          type: string
          enum: [eda, analysis, model, full]
          default: full
    AnalysisStep:
      type: object
      required: [step_id, name, status, order]
      properties:
        step_id:
          type: string
        name:
          type: string
        tool_name:
          type: [string, "null"]
        tool_version:
          type: [string, "null"]
        status:
          type: string
          enum: [pending, running, succeeded, failed, skipped, blocked, cancelled]
        order:
          type: integer
          minimum: 1
        started_at:
          type: [string, "null"]
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
        artifact_ids:
          type: array
          items:
            type: string
        error:
          anyOf:
            - $ref: "#/components/schemas/ErrorDetail"
            - type: "null"
    AnalysisRun:
      type: object
      required:
        - run_id
        - project_id
        - dataset_version_id
        - analysis_spec_id
        - status
        - progress
        - steps
        - created_at
      properties:
        run_id:
          type: string
          example: run_01JABC
        project_id:
          type: string
        dataset_version_id:
          type: string
        analysis_spec_id:
          type: string
        source_run_id:
          type: [string, "null"]
        run_kind:
          type: string
          enum: [eda, analysis, model, full]
        status:
          type: string
          enum: [queued, running, blocked, succeeded, failed, cancelled]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        current_step_id:
          type: [string, "null"]
        steps:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisStep"
        environment:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        started_at:
          type: [string, "null"]
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
        error:
          anyOf:
            - $ref: "#/components/schemas/ErrorDetail"
            - type: "null"
    AnalysisRunPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AnalysisRun"
...
    ReportExportRequest:
      type: object
      additionalProperties: false
      required: [run_id, format]
      properties:
        run_id:
          type: string
        format:
          type: string
          enum: [html, notebook, cleaned_data, manifest]
        claim_ids:
          type: array
          items:
            type: string
        include_code:
          type: boolean
          default: true
        include_evidence:
          type: boolean
          default: true
        data_format:
          type: [string, "null"]
          enum: [csv, parquet, null]
...
    Job:
      type: object
      required:
        - job_id
        - kind
        - status
        - progress
        - created_at
        - updated_at
      properties:
        job_id:
          type: string
          example: job_01JABC
        kind:
          type: string
          enum:
            - dataset_ingestion
            - version_comparison
            - quality_scan
            - cleaning_preview
            - cleaning_execute
            - analysis_run
            - report_export
        status:
          type: string
          enum: [queued, running, cancelling, cancelled, blocked, succeeded, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        current_step:
          type: [string, "null"]
        resource_type:
          type: [string, "null"]
        resource_id:
          type: [string, "null"]
        retry_after_ms:
          type: [integer, "null"]
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        error:
          anyOf:
            - $ref: "#/components/schemas/ErrorDetail"
            - type: "null"
    PageMeta:
      type: object
      required: [page, page_size, total, has_more]
      properties:
        page:
          type: integer
          minimum: 1
        page_size:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        has_more:
          type: boolean
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message, request_id, retryable, details]
      properties:
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]+$"
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true